backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pycryptodome==3.19.1
